[feat] Lottie Loading 추가
