+++ next-netflix/package.json
@@ -17,6 +17,7 @@
     "axios": "^1.4.0",
     "eslint": "8.40.0",
     "eslint-config-next": "^13.4.1",
+    "lottie-react": "^2.4.0",
     "next": "^13.4.1",
     "next-navigation": "^1.0.6",
     "postcss": "8.4.23",
